# streaming store-gateway: avoid copying postings when hashing them