@dimitarvdimitrov 2022-12-20
These are the results of running `BenchmarkCanonicalPostingsKey`

```
name                                      old time/op    new time/op    delta
CanonicalPostingsKey/10000_postings-10       202µs ± 0%     107µs ± 2%   -47.12%  (p=0.008 n=5+5)
CanonicalPostingsKey/1000000_postings-10    20.0ms ± 0%    10.5ms ± 0%   -47.48%  (p=0.008 n=5+5)
CanonicalPostingsKey/100000_postings-10     2.00ms ± 0%    1.05ms ± 0%   -47.70%  (p=0.008 n=5+5)
CanonicalPostingsKey/100_postings-10        2.46µs ± 3%    1.27µs ± 3%   -48.13%  (p=0.008 n=5+5)
CanonicalPostingsKey/1000_postings-10       20.8µs ± 0%    10.6µs ± 0%   -48.87%  (p=0.016 n=5+4)
CanonicalPostingsKey/10_postings-10          497ns ± 1%     251ns ± 2%   -49.45%  (p=0.008 n=5+5)

name                                      old alloc/op   new alloc/op   delta
CanonicalPostingsKey/10_postings-10           720B ± 0%       96B ± 0%   -86.67%  (p=0.008 n=5+5)
CanonicalPostingsKey/100_postings-10        1.54kB ± 0%    0.10kB ± 0%   -93.75%  (p=0.008 n=5+5)
CanonicalPostingsKey/1000_postings-10       8.83kB ± 0%    0.10kB ± 0%   -98.91%  (p=0.008 n=5+5)
CanonicalPostingsKey/10000_postings-10      82.6kB ± 0%     0.1kB ± 0%   -99.88%  (p=0.008 n=5+5)
CanonicalPostingsKey/100000_postings-10      803kB ± 0%       0kB ± 0%   -99.99%  (p=0.008 n=5+5)
CanonicalPostingsKey/1000000_postings-10    8.00MB ± 0%    0.00MB ± 0%  -100.00%  (p=0.008 n=5+5)

name                                      old allocs/op  new allocs/op  delta
CanonicalPostingsKey/10_postings-10           6.00 ± 0%      2.00 ± 0%   -66.67%  (p=0.008 n=5+5)
CanonicalPostingsKey/100_postings-10          6.00 ± 0%      2.00 ± 0%   -66.67%  (p=0.008 n=5+5)
CanonicalPostingsKey/1000_postings-10         6.00 ± 0%      2.00 ± 0%   -66.67%  (p=0.008 n=5+5)
CanonicalPostingsKey/10000_postings-10        6.00 ± 0%      2.00 ± 0%   -66.67%  (p=0.008 n=5+5)
CanonicalPostingsKey/100000_postings-10       6.00 ± 0%      2.00 ± 0%   -66.67%  (p=0.008 n=5+5)
CanonicalPostingsKey/1000000_postings-10      6.00 ± 0%      2.00 ± 0%   -66.67%  (p=0.008 n=5+5)
```