@pracucci 2022-12-20
I'm wondering if we could also have a simple unit test on `CanonicalPostingsKey()`. Like, same input, same hash. Different input, different hash.